test(chat_presenter): split cancel pipeline tests into submodule
Structural check enforces a 1000-line hard limit. chat_presenter_tests.rs
had grown to 1104 lines, failing the Lint job on PR #176.

Extract the cancel-pipeline block (RecordingChatService, the two
StopStreaming tests, and MockAppSettingsService) into a sibling file
included via #[path] mod cancel_tests inside the existing tests module.
This keeps all test items reachable through super::* while bringing
every file back under the 1000-line cap.

chat_presenter_tests.rs: 888 lines.
chat_presenter_cancel_tests.rs: 222 lines.

Verified locally: cargo fmt, cargo clippy --all-targets -D warnings,
and the chat_presenter test suite (15 tests, all pass including both
cancel_tests).
